Add flexible database connection management
The database process pool now always has a fixed size (per site, in the
`db_max_connections` config key, defaults to 10), but the pool workers
start in disconnected state. When the workers are being used they start
a connection, and, after a timeout (currently hardcoded to 60 seconds),
they drop their connection again.

For this change, a poolboy PR was made to be able to give preference to
hot worker processes: https://github.com/devinus/poolboy/pull/50